requirements.txt
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.4.0
beautifulsoup4==4.13.3
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.4.0
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
filelock==3.15.4
fsspec==2024.6.1
google-ai-generativelanguage==0.1.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.19.0
google-crc32c==1.5.0
google-generativeai==0.1.0rc1
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.26.3
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json_repair==0.30.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
openai==1.42.0
packaging==24.2
pillow==10.4.0
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rapidfuzz==3.9.7
regex==2024.9.11
replicate==1.0.4
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.2
threadpoolctl==3.5.0
tiktoken==0.7.0
timeout-decorator==0.5.0
tokenizers==0.20.3
torch==2.4.0
tqdm==4.66.5
transformers==4.46.3
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
vertexai==1.71.1
